# vote_demo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Vote API",
            "email": "yllmis"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/login": {
            "post": {
                "description": "执行用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "执行用户登录",
                "parameters": [
                    {
                        "description": "login User",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "执行用户退出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "执行用户退出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        },
        "/vote": {
            "get": {
                "description": "获取投票信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vote"
                ],
                "summary": "获取投票信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "vote ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "logic.User": {
            "type": "object",
            "properties": {
                "captcha_id": {
                    "description": "新增：验证码ID",
                    "type": "string"
                },
                "captcha_value": {
                    "description": "新增：用户输入的验证码",
                    "type": "string"
                },
                "name": {
                    "description": "注意这里的 tag 要和表单中input的 name 属性对应",
                    "type": "string"
                },
                "password": {
                    "description": "注意，若password首字母小写，则无法绑定到，因为反射只能访问导出字段",
                    "type": "string"
                }
            }
        },
        "tools.ECode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}